feat: hide system components when switch enabled
